requirements.txt
ccxt>=4.3.75
httpx>=0.27.0
pydantic>=2.9.0
python-dotenv>=1.0.1
git+https://github.com/hyperliquid-dex/hyperliquid-python-sdk.git#egg=hyperliquid-python-sdk
